vocal-lingua-backend/requirements.txt
# Gabsther Backend Dependencies
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
Pillow>=10.0
psycopg2-binary>=2.9
django-filter>=23.5
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
groq>=1.1
